[FLINK-8742][docs] Move docs generator annotations to flink-annotations
This closes #5821.
This closes #5787.
