+++ flink-annotations/src/main/java/org/apache/flink/annotation/docs/ConfigGroups.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.configuration;
+package org.apache.flink.annotation.docs;
 
 import org.apache.flink.annotation.Internal;
 
@@ -26,8 +26,8 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Annotation used on classes containing {@link ConfigOption}s that enables the separation of options into different
- * tables based on key prefixes. A {@link ConfigOption} is assigned to a {@link ConfigGroup} if the option key matches
+ * Annotation used on classes containing config optionss that enables the separation of options into different
+ * tables based on key prefixes. A config option is assigned to a {@link ConfigGroup} if the option key matches
  * the group prefix. If a key matches multiple prefixes the longest matching prefix takes priority. An option is never
  * assigned to multiple groups. Options that don't match any group are implicitly added to a default group.
  */
